Keyboard navigation feature refinement.
Fixed a lot of edge cases bugs that caused errors.

Now the highlight is only navigated within the currenlty filtered
images.

Keyboard navigation is made active solely when the search input is out
of focus. This avoids mixing cursor movement in the input field, with
highlight movement amongst the images. Tab changes focus from the input
field to the first image.